microbit: Serial redesign
The previous design of the serial API was too specific with regards to type and
didn't align with the asynchronous nature of the DAL.

This revised version fixes these two issues.

A number of methods have been revised and the behaviour is now to block the current
fiber and wait for an event, allowing other fibers to continue execution.

There is also a mutex lock on the serial bus so only one fiber can use it at a time.