microbit: MicroBitSerial redesign
This implementation uses a circular buffer for both transmission and reception of data.

There are a number of new methods:

	* redirect - which dynamically redirects the output of the serial module
	* eventOn and eventAfter - which fire events based on delimeters, or x number of characters.
	* operations on the circular buffers.

A main difference is the ability to select the mode you want to perform your operation in:

	* ASYNC - reads directly from the circular buffer, returning immediately.
	* SYNC_SPINWAIT - spins whilst reading from the circular buffer.
	* SYNC_SLEEP - blocks the current fiber, but doesn't lock out the processor,
		       whilst reading from the circular buffer.

This should support a wide variety of operations. To see more specific details of the different methods,
it is advised to read the doxygen comments for each method.