microbit: changes to the default servo range
Previously the range of the servo API was incorrect and would only
translate to a range 0 degrees - 90 degrees in the real world.

This patch updates the #defines used in the servo api default
calculation to calibrate for the intended range of 0 - 180 degrees.

This was all due to working from a specification that applied to a
subset of servos.