microbit-dal: initial refactor to MicroBitAccelerometer
This refactor removes reliance on the scheduler. It allows the
accelerometer to be used even when the scheduler is not active,
and no idle thread is available for updating the sample used by
the accelerometer.

This commit also provides an optimisation through the lazy
addition of the accelerometer to the idle thread.