microbit: Added EventModel interface.
The microbit-dal runtime uses asynchronous events as a mechanism to decouple
user code from interrupt context and to provide an approach to asynchronous
programming on the micro:bit.

Not all languages may wish to use the default MessageBus service, and instead
prefer to roll their own event model. This patch adds a level of indirection
in the form of an EventModel interface that defines an API that allows any
EventModel to consume events from the microbit-dal runtime with whichever
implementaiton they choose. Such event models should subclass the
EventModel interface, and implement some (or none) of the defined functions:

 - send()
 - add()
 - remove()
 - elemenAt()

All uses of MicroBitMessageBus updated to use EventModel, and all event
listeners within the runtime now attmept to attach to a default service -
EventModel::defaultEventBus (which may or may not be present).