microbit: moved MicroBitStorage to a Key Value Store
MicroBitStorage used to rely on a static struct as a memory
map when writing to flash "MicroBitConfigBlock", this solution
was constraining, and not fit for general purpose.

Whilst modularising the runtime, a cyclic dependency between
MicroBitStorage and MicroBitCompass was discovered, which meant
these two class could not be fully decoupled.

This version of MicroBitStorage is a Key Value Store, and allows
3 base operations:

	* put - add a new KeyValuePair to flash
	* get - retreive a KeyValuePair from flash
	* remove - delete a KeyValuePair from flash

An instance of MicroBitStorage is also now a part of uBit (uBit.storage)
and this is now passed as a parameter to the constructors of BLEManager
and compass, both of which require the persistence of data.

There are also overloads for the constructors of MicroBitBLEManager
and MicroBitCompass, which do not require an instance of
MicroBitStorage to be passed.