microbit: Updated MicroBitSystemTimer to use dynamically allocated Ticker
MicroBitSystemTimer uses a statically allocated Ticker object as a timer
source. This both incurs a static memory overhead when the timer is never
initialised, and causes temporal dependencies for applicaitons that
initialise the timer from within a static context.

This patch simply changes the implementaiton of MicroBitSystemTimer to
dynamically allocate the Ticker object on the heap when it is first used to
address both of these issues.

The patch also cleans up an unecessary call to system_timer_init() in
MicroBit::init().