microbit: MicroBitThermometer lazy addition to fiber components
Previously, MicroBitThermometer was always added to fiber_components,
even in the case where it wasn't actually in use.

This commit introduces a similar behaviour to that introduced to the compass
and accelerometer classes in previous commits, where these components
are only added to fiber_components when in use.

Similarly, MicroBitThermometer now has an exposed updateSample() member
function that can be used to manually drive temperature updates, if
the fiber scheduler is not in use.