microbit: fixed overload ambiguity in MicroBitStorage
MicroBitStorage accepts both ManagedString and char * variants, this
caused ambiguity at compile time, and meant that the ManagedString
variant was never selected when given character literals. This meant
that the only way to access the ManagedString variant, was to explicitly
wrap the character literals.

This applies to:
	* put
	* get
	* remove

The fix was to apply a const prefix to all member functions accepting
a char *.