microbit: BUGFIX for MicroBitStorage
There was a bug with Bluetooth pairing. If a user were to bond with the
micro:bit, and the KeyValueStore region had not been configured, the
class would attempt to write to flash in interrupt context, which
conflicted with soft device.

The fix is to do this initialisation in the constructor for
MicroBitStorage.