microbit-dal: Added send modes to the UART service. [#146]
Added a blocking send to the UART service for the case when the bytes
to transmit is greater than the size of the transmission buffer.

The characteristics have now diverged from the BLE_API’s
implementation, and are now more semantic.

Send now returns an error code if there is no device connected, or
indications have not been enabled.

Debug statements have also been removed.