microbit-dal: Modifications to MicroBitSerial for subclassing
If a user were to subclass MicroBitSerial, a number of functions would
be unavailable. This commit moves certain functions to protected rather
than private.

If one were to subclass MicroBitSerial and MicroBitComponent, there
would be a naming collision with MicroBitComponents’s status variable.
This commit also renames status to serial_status