microbit-dal: Added JMXSerial[.h,.cpp]
Added JMXSerial which introduces the capability to respond to file
system requests from the interface chip over the Serial line.

JMXSerial is a subclass of MicroBitSerial. To enable filesystem
interaction, an instance of JMXSerial needs to be created within the
first few seconds of power on. (This capability will only work with a
specific revision of the interface firmware)

There are a few new configuration options added to MicroBitConfig.h:

* MICROBIT_IF_CHIP_FS - Enables the file system if set.
* MICROBIT_IF_CHIP_PRESERVE  - Enables page preservation during
flashing.
* MICROBIT_IF_PRES_1 - the first page to preserve.
* MICROBIT_IF_PRES_2 - the second page to preserve.