microbit-dal: removed ManagedString APIs for Eddystone Uid
ManagedString should only be used for string based data.
The APIs removed by this commit presumed that this was not the case.
ManagedString implicitly contains error checking, which makes it
unsuitable for sequences of bytes that may contain null termination
characters (0x00).

Removing these APIs for now was the best option until a future time
when ManagedBuffer can be used within the runtime.