microbit-dal: DynamicPWM now has a static period member variable
Previously DynamicPWM had a per instance member variable that was not shared across all instances of DyanmicPWM which is not true to the hardware, that shares a period across all pwm channels.

This commit adds a static member variable, and sets the default to 20000 us - a default set at the mbed layer.