NEW UX: update compass calibration to fill screen
As per #288, this commit introduces a full screen compass calibration
that collects more data points than the older regime. In conjunction
with the newer algorithm developed by @finneyj this improves compass
accuracy after a calibration and avoids the common bad-case of heavy
skew towards one direction.

The new routine asks the user to "Tilt to fill screen" instead of "Draw
a circle". The addition of "Tilt" in the instruction and the
uncontestability of 'full screen' make this an easier step for a user to
complete with certainty.

This commit also removes the older mechanisms from the prototype and
cleans up the unit scaling to ensure correct SI units. This required a
modification to the hill descent to ensure it converges in sane time.