Make calibration routines static, separate from UX
This commit makes the calibration routines independent of the UX for
calibrating the compass, so that it's not necessary to create a
MicroBitCompassCalibrator instance before doing a calibration.

This in turns lets the calibration UX be programmed using different
button and display primatives while the calibration code can easily
be shared.