Compass calibration: collect data while scrolling
Two issues with the current compass calibration 'game' are that
* The instructions are only shown once at the beginning of the game
* The instructions cannot be skipped and block the user

This commit addresses these issues by moving the text scroll inside the
main data collection loop. We collect data in the background from the
moment the message starts scrolling and if we reach the threshold number
of points collected (currently 15) then the message stops.

If the calibration is not completed inside the allocated time period
(here 30s) then the scrolling message repeats.

This should allow advanced users to calibrate quickly, and gives users
who didn't read the instructions a second (and third, and fourth...)
chance to understand.

We still have to address the localisation issue eventually.