UX Change: Change compass calibration routine
This commit changes the user experience for calibrating the compass. It
does *not* change the algorithm at all, and so issues already known are
still present (#288 for example).

However, we can expect better results from having more calibration
points and an easier to understand user-action.

The new algorithm asks the user "Tilt to fill screen" and rather than
taking 12 points, looks for 25 points, represented by each pixel on the
screen. By tilting the micro:bit the user will fill the whole screen and
the calibration will run.

This commit also merges the 'help' message display loop with the main
calibration loop in order to
* Allow the 'help' message to be redisplayed after a timeout (default
30s)
* Allow the user to start collecting data points immediately. You can
start tilting as soon as the message starts, and if you reach a number
of points (default 15) then the message will stop to show you the
calibration screen.

This should annoy advanced users less as well as helping new users do
the right thing.