This is in a better state.
I've refactored the protocol to include time at origin, and ttl at origin. From this we can calculate the time taken to traverse the network, and wake up at precisely the same time as the origin node.

Unfortunately, I'm now facing problems with the way we check for tx and sleep callbacks, it's causing jitter on the scope. I'll have to think about refactoring or reducing the number of callbacks.