changed period index and added (commented out) group concencusing
Periods in the periods array are now mutiples of one another so that conflicting periods is impossible.

This combined with a new concensus approach will enable no-conflict mesh networks.

The new concensus approach which doesn't quite work and requires debugging. It uses a local speed variable to determine if it itself needs to speed up or doesn't need to go as fast. If it would like to change speed, it puts a request in the next frame formed as a proposal. Other nodes can then repeat this packet as an ack, or ignore to reject the proposal. As long as one device acks, the mesh can still operate effectively.

The assumed approach doesn't work yet, and requires testing at a later date, for now lets stick to 20ms